refactor out keys_iterator
