+++ src/lib.rs
@@ -222,14 +222,16 @@ pub mod codes_handle;
 pub mod codes_index;
 pub mod codes_nearest;
 pub mod errors;
+pub mod keys_iterator;
 mod intermediate_bindings;
 mod pointer_guard;
 
-pub use codes_handle::{CodesHandle, Key, KeyType, KeyedMessage, KeysIteratorFlags, ProductKind};
+pub use codes_handle::{CodesHandle, Key, KeyType, KeyedMessage, ProductKind};
 #[cfg(feature = "experimental_index")]
 #[cfg_attr(docsrs, doc(cfg(feature = "experimental_index")))]
 pub use codes_index::CodesIndex;
 pub use codes_nearest::{CodesNearest, NearestGridpoint};
+pub use keys_iterator::{KeysIterator, KeysIteratorFlags};
 pub use errors::CodesError;
 pub use fallible_iterator::{FallibleIterator, IntoFallibleIterator};
 pub use fallible_streaming_iterator::FallibleStreamingIterator;
